+++ x/btcstaking/types/events.go
@@ -1,9 +1,23 @@
 package types
 
+import (
+	bbn "github.com/babylonchain/babylon/types"
+)
+
 func NewEventPowerDistUpdateWithBTCDel(ev *EventBTCDelegationStateUpdate) *EventPowerDistUpdate {
 	return &EventPowerDistUpdate{
 		Ev: &EventPowerDistUpdate_BtcDelStateUpdate{
 			BtcDelStateUpdate: ev,
+		},
+	}
+}
+
+func NewEventPowerDistUpdateWithSlashedFP(fpBTCPK *bbn.BIP340PubKey) *EventPowerDistUpdate {
+	return &EventPowerDistUpdate{
+		Ev: &EventPowerDistUpdate_SlashedFp{
+			SlashedFp: &EventPowerDistUpdate_EventSlashedFinalityProvider{
+				Pk: fpBTCPK,
+			},
 		},
 	}
 }
